requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==2.0.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
opencv-python-headless
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
